Transferred all query functions to one file
